pyproject.toml
[project]
name = "parserx"
version = "0.1.0"
description = "High-fidelity document parsing tool for knowledge bases, retrieval, and analysis"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymupdf>=1.25.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "openai>=1.0",
    "pillow>=10.0",
    "numpy>=1.26",
]

[project.optional-dependencies]
docx = ["docling>=2.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[project.scripts]
parserx = "parserx.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
